uncast the mask

+++ bit_manipulation/1-print_binary.c
@@ -9,12 +9,12 @@ void print_binary(unsigned long int n)
 {
 	int i;
 	unsigned long int mask;
-	int num_bits = sizeof(unsigned long int) * 8;
+	int num_bits = sizeof(int) * 8;
 	int significant_bits_found = 0;
 
 	for (i = num_bits - 1; i >= 0; i--)
 	{
-		mask = (unsigned long int)1 << i;
+		mask = 1 << i;
 		if (n & mask)
 		{
 			_putchar(0 + '1');
